pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supportmind"
version = "0.1.0"
description = "Self-Learning AI Support Intelligence System"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ai", "support", "rag", "knowledge-base", "self-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
api = ["fastapi>=0.100.0", "uvicorn>=0.23.0", "pydantic>=2.0.0"]
dashboard = ["streamlit>=1.25.0", "plotly>=5.15.0"]
gpu = [
  "faiss-gpu-cu12; platform_system=='Linux' and platform_machine=='x86_64'"
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.4.0",
    "pre-commit>=3.3.0",
]
all = [
  "fastapi>=0.100.0", "uvicorn>=0.23.0", "pydantic>=2.0.0",
  "streamlit>=1.25.0", "plotly>=5.15.0",
  "pytest>=7.4.0", "pytest-cov>=4.1.0", "pytest-asyncio>=0.21.0",
  "black>=23.7.0", "isort>=5.12.0", "flake8>=6.1.0", "mypy>=1.4.0", "pre-commit>=3.3.0"
]

[project.scripts]
supportmind = "supportmind.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/supportmind"
Documentation = "https://github.com/yourusername/supportmind#readme"
Repository = "https://github.com/yourusername/supportmind"
Issues = "https://github.com/yourusername/supportmind/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["supportmind*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=supportmind --cov-report=term-missing"
